Remove game plugin and move code to editor assembly
