+++ UnitTests.Editor.csproj
@@ -59,6 +59,9 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Cache\AssemblyInfo.cs" />
+    <Compile Include="Source\Editor\Assert.cs" />
+    <Compile Include="Source\Editor\ExampleClass.cs" />
+    <Compile Include="Source\Editor\TestAttributes.cs" />
     <Compile Include="Source\Editor\TestRunner.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
